Type:feat(17831 new demand)
Descripton:Voice Note prompts for forced shutdown / restart / logout during recording.

RootCause:Null

Solution:How do you fix the bug

+++ src/views/mainwindow.cpp
@@ -902,7 +902,11 @@ void MyMainWindow::deleteNoteShortcut()
     {
         return;
     }
-    m_mainPage->deleteFolder();
+
+    if (!Intancer::get_Intancer()->getRecodingFlag()
+            && !Intancer::get_Intancer()->getVoiceToTextFlag()) {
+        m_mainPage->deleteFolder();
+    }
 }
 void MyMainWindow::VoiceNotesPlayShortcut()
 {
